@@ -16,16 +16,17 @@ namespace Fragen\Translations_Updater;
  * @package Fragen\Translations_Updater
  */
 class Language_Pack_API {
+
 	use API;
 
 	/**
 	 * Constructor.
 	 *
-	 * @param object $type
+	 * @param \stdClass $config Repo config data.
 	 */
-	public function __construct( $type ) {
-		$this->type     = $type;
-		$this->response = $this->get_repo_cache();
+	public function __construct( $config ) {
+		$this->repo     = $config;
+		$this->response = $this->get_repo_cache( $config->slug );
 	}
 
 	/**
@@ -37,54 +38,62 @@ class Language_Pack_API {
 	 */
 	public function get_language_pack( $headers ) {
 		$response = ! empty( $this->response['languages'] ) ? $this->response['languages'] : false;
-		$type     = explode( '_', $this->type->type );
 
 		if ( ! $response ) {
-			$response = $this->get_language_pack_json( $type[0], $headers, $response );
+			$response = $this->get_language_pack_json( $this->repo->git, $headers );
 
 			if ( $response ) {
 				foreach ( $response as $locale ) {
-					$package = $this->process_language_pack_package( $type[0], $locale, $headers );
+					$package = $this->process_language_pack_package( $this->repo->git, $locale, $headers );
 
 					$response->{$locale->language}->package = $package;
-					$response->{$locale->language}->type    = $type[1];
-					$response->{$locale->language}->version = $this->type->local_version;
+					$response->{$locale->language}->type    = $this->repo->type;
+					$response->{$locale->language}->version = $this->repo->version;
 				}
-
-				$this->set_repo_cache( 'languages', $response );
+				$this->set_repo_cache( 'languages', $response, $this->repo->slug );
 			} else {
 				return false;
 			}
 		}
+		$this->repo->language_packs = $response;
 
-		$this->type->language_packs = $response;
+		return $this->repo;
 	}
 
 	/**
 	 * Get language-pack.json from appropriate host.
 	 *
-	 * @param string $type     ( github|bitbucket|gitlab )
+	 * @param string $git ( github|bitbucket|gitlab|gitea )
 	 * @param array  $headers
-	 * @param mixed  $response API response.
 	 *
-	 * @return array|bool|mixed|object
+	 * @return array|bool|mixed|object $response API response object.
 	 */
-	private function get_language_pack_json( $type, $headers, $response ) {
-		switch ( $type ) {
+	private function get_language_pack_json( $git, $headers ) {
+		switch ( $git ) {
 			case 'github':
 				$response = $this->api( '/repos/' . $headers['owner'] . '/' . $headers['repo'] . '/contents/language-pack.json' );
-				$contents = base64_decode( $response->content );
-				$response = json_decode( $contents );
+				$response = isset( $response->content )
+					? json_decode( base64_decode( $response->content ) )
+					: null;
 				break;
 			case 'bitbucket':
 				$response = $this->api( '/1.0/repositories/' . $headers['owner'] . '/' . $headers['repo'] . '/src/master/language-pack.json' );
-				$response = json_decode( $response->data );
+				$response = isset( $response->data )
+					? json_decode( $response->data )
+					: null;
 				break;
 			case 'gitlab':
 				$id       = urlencode( $headers['owner'] . '/' . $headers['repo'] );
 				$response = $this->api( '/projects/' . $id . '/repository/files/language-pack.json' );
-				$contents = base64_decode( $response->content );
-				$response = json_decode( $contents );
+				$response = isset( $response->content )
+					? json_decode( base64_decode( $response->content ) )
+					: null;
+				break;
+			case 'gitea':
+				$response = $this->api( '/repos/' . $headers['owner'] . '/' . $headers['repo'] . '/raw/master/language-pack.json' );
+				$response = isset( $response->content )
+					? json_decode( base64_decode( $response->content ) )
+					: null;
 				break;
 		}
 
@@ -98,31 +107,35 @@ class Language_Pack_API {
 	/**
 	 * Process $package for update transient.
 	 *
-	 * @param string $type ( github|bitbucket|gitlab )
+	 * @param string $git ( github|bitbucket|gitlab|gitea )
 	 * @param string $locale
 	 * @param array  $headers
 	 *
 	 * @return array|null|string
 	 */
-	private function process_language_pack_package( $type, $locale, $headers ) {
+	private function process_language_pack_package( $git, $locale, $headers ) {
 		$package = null;
-		switch ( $type ) {
+		switch ( $git ) {
 			case 'github':
-				$package = array( 'https://github.com', $headers['owner'], $headers['repo'], 'blob/master' );
+				$package = [ 'https://github.com', $headers['owner'], $headers['repo'], 'blob/master' ];
 				$package = implode( '/', $package ) . $locale->package;
-				$package = add_query_arg( array( 'raw' => 'true' ), $package );
+				$package = add_query_arg( [ 'raw' => 'true' ], $package );
 				break;
 			case 'bitbucket':
-				$package = array( 'https://bitbucket.org', $headers['owner'], $headers['repo'], 'raw/master' );
+				$package = [ 'https://bitbucket.org', $headers['owner'], $headers['repo'], 'raw/master' ];
 				$package = implode( '/', $package ) . $locale->package;
 				break;
 			case 'gitlab':
-				$package = array( 'https://gitlab.com', $headers['owner'], $headers['repo'], 'raw/master' );
+				$package = [ 'https://gitlab.com', $headers['owner'], $headers['repo'], 'raw/master' ];
 				$package = implode( '/', $package ) . $locale->package;
+				break;
+			case 'gitea':
+				// TODO: make sure this works
+				$package = [ $headers['uri'], 'raw/master' ];
+				$package = implode( '/', $package ) . $local->package;
 				break;
 		}
 
 		return $package;
 	}
-
 }
