@@ -39,47 +39,35 @@ trait API {
 	protected $response = array();
 
 	/**
-	 * Adds custom user agent for Translations Updater.
-	 *
-	 * @param array  $args Existing HTTP Request arguments.
-	 * @param string $url  URL being passed.
-	 *
-	 * @return array Amended HTTP Request arguments.
-	 */
-	public static function http_request_args( $args, $url ) {
-		$args['sslverify'] = true;
-		if ( false === strpos( $args['user-agent'], 'Translations Updater' ) ) {
-			$args['user-agent'] .= '; Translations Updater - https://github.com/afragen/translations-updater';
-		}
-
-		return $args;
-	}
-
-	/**
 	 * Return repo data for API calls.
 	 *
 	 * @return array
 	 */
 	protected function return_repo_type() {
-		$type        = explode( '_', $this->type->type );
 		$arr         = array();
-		$arr['type'] = $type[1];
+		$arr['type'] = $this->repo->type;
 
-		switch ( $type[0] ) {
+		switch ( $this->repo->git ) {
 			case 'github':
-				$arr['repo']          = 'github';
+				$arr['git']           = 'github';
 				$arr['base_uri']      = 'https://api.github.com';
 				$arr['base_download'] = 'https://github.com';
 				break;
 			case 'bitbucket':
-				$arr['repo']          = 'bitbucket';
+				$arr['git']           = 'bitbucket';
 				$arr['base_uri']      = 'https://bitbucket.org/api';
 				$arr['base_download'] = 'https://bitbucket.org';
 				break;
 			case 'gitlab':
-				$arr['repo']          = 'gitlab';
+				$arr['git']           = 'gitlab';
 				$arr['base_uri']      = 'https://gitlab.com/api/v4';
 				$arr['base_download'] = 'https://gitlab.com';
+				break;
+			case 'gitea':
+				$arr['git'] = 'gitea';
+				// TODO: make sure this works
+				$arr['base_uri']      = $this->repo->languages . '/api/v1';
+				$arr['base_download'] = $this->repo->languages;
 				break;
 		}
 
@@ -88,15 +76,12 @@ trait API {
 
 	/**
 	 * Call the API and return a json decoded body.
-	 * Create error messages.
 	 *
 	 * @param string $url
 	 *
 	 * @return boolean|\stdClass
 	 */
 	protected function api( $url ) {
-		add_filter( 'http_request_args', array( &$this, 'http_request_args' ), 10, 2 );
-
 		$response = wp_remote_get( $this->get_api_url( $url ) );
 
 		if ( is_wp_error( $response ) ) {
@@ -118,9 +103,10 @@ trait API {
 	protected function get_api_url( $endpoint ) {
 		$type = $this->return_repo_type();
 
-		switch ( $type['repo'] ) {
+		switch ( $type['git'] ) {
 			case 'github':
 			case 'bitbucket':
+			case 'gitea':
 				break;
 			case 'gitlab':
 				$endpoint = add_query_arg( 'ref', 'master', $endpoint );
@@ -151,12 +137,12 @@ trait API {
 	 */
 	protected function get_repo_cache( $repo = false ) {
 		if ( ! $repo ) {
-			$repo = isset( $this->type->repo ) ? $this->type->repo : 'tu';
+			$repo = isset( $this->type->slug ) ? $this->type->slug : 'tu';
 		}
 		$cache_key = 'tu-' . md5( $repo );
 		$cache     = get_site_option( $cache_key );
 
-		if ( empty( $cache['timeout'] ) || current_time( 'timestamp' ) > $cache['timeout'] ) {
+		if ( empty( $cache['timeout'] ) || time() > $cache['timeout'] ) {
 			return false;
 		}
 
@@ -174,17 +160,16 @@ trait API {
 	 */
 	protected function set_repo_cache( $id, $response, $repo = false ) {
 		if ( ! $repo ) {
-			$repo = isset( $this->type->repo ) ? $this->type->repo : 'tu';
+			$repo = isset( $this->type->slug ) ? $this->type->slug : 'tu';
 		}
 		$cache_key = 'tu-' . md5( $repo );
 		$timeout   = '+' . self::$hours . ' hours';
 
-		$this->response['timeout'] = strtotime( $timeout, current_time( 'timestamp' ) );
+		$this->response['timeout'] = strtotime( $timeout );
 		$this->response[ $id ]     = $response;
 
 		update_site_option( $cache_key, $this->response );
 
 		return true;
 	}
-
 }
