@@ -23,6 +23,7 @@ if ( ! defined( 'WPINC' ) ) {
  * @package Fragen\Translations_Updater
  */
 class Language_Pack {
+
 	use Base;
 
 	/**
@@ -35,16 +36,15 @@ class Language_Pack {
 	/**
 	 * Language_Pack constructor.
 	 *
-	 * @param Plugin|Theme                                   $repo Plugin/Theme object.
+	 * @param \stdClass                                      $config
 	 * @param \Fragen\Translations_Updater\Language_Pack_API $api  Language_Pack_API object.
 	 */
-	public function __construct( $repo, Language_Pack_API $api ) {
-		if ( null === $repo->languages ) {
+	public function __construct( $config, Language_Pack_API $api ) {
+		if ( null === $config->languages ) {
 			return;
 		}
 
-		$this->add_headers();
-		$this->repo     = $repo;
+		$this->repo     = $config;
 		$this->repo_api = $api;
 	}
 
@@ -52,11 +52,12 @@ class Language_Pack {
 	 * Do the Language Pack integration.
 	 */
 	public function run() {
-		$headers = $this->parse_header_uri( $this->repo->languages );
-		$this->repo_api->get_language_pack( $headers );
+		$headers                     = $this->parse_header_uri( $this->repo->languages );
+		$repo                        = $this->repo_api->get_language_pack( $headers );
+		$this->config[ $repo->slug ] = $repo;
 
-		add_filter( 'pre_set_site_transient_update_plugins', array( &$this, 'pre_set_site_transient' ) );
-		add_filter( 'pre_set_site_transient_update_themes', array( &$this, 'pre_set_site_transient' ) );
+		add_filter( 'pre_set_site_transient_update_plugins', [ $this, 'pre_set_site_transient' ] );
+		add_filter( 'pre_set_site_transient_update_themes', [ $this, 'pre_set_site_transient' ] );
 	}
 
 	/**
@@ -76,16 +77,15 @@ class Language_Pack {
 		}
 
 		if ( 'pre_set_site_transient_update_plugins' === current_filter() ) {
-			$repos        = \Fragen\Singleton::get_instance( 'Plugin', $this )->get_plugin_configs();
 			$translations = wp_get_installed_translations( 'plugins' );
 		}
 		if ( 'pre_set_site_transient_update_themes' === current_filter() ) {
-			$repos        = \Fragen\Singleton::get_instance( 'Theme', $this )->get_theme_configs();
 			$translations = wp_get_installed_translations( 'themes' );
 		}
 
 		$repos = array_filter(
-			$repos, function( $e ) {
+			$this->config,
+			function ( $e ) {
 				return isset( $e->language_packs );
 			}
 		);
@@ -95,8 +95,8 @@ class Language_Pack {
 				$lang_pack_mod   = isset( $repo->language_packs->$locale )
 					? strtotime( $repo->language_packs->$locale->updated )
 					: 0;
-				$translation_mod = isset( $translations[ $repo->repo ][ $locale ] )
-					? strtotime( $translations[ $repo->repo ][ $locale ]['PO-Revision-Date'] )
+				$translation_mod = isset( $translations[ $repo->slug ][ $locale ] )
+					? strtotime( $translations[ $repo->slug ][ $locale ]['PO-Revision-Date'] )
 					: 0;
 				if ( $lang_pack_mod > $translation_mod ) {
 					$transient->translations[] = (array) $repo->language_packs->$locale;
@@ -108,5 +108,4 @@ class Language_Pack {
 
 		return $transient;
 	}
-
 }
